requirements.txt
Flask==3.1.0
boto3==1.34.0
Werkzeug==3.1.0
Flask-Limiter==3.5.0
Flask-WTF==1.2.2
python-dotenv==1.0.1
PyJWT==2.7.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1